02_financial_agent_with_phidata/requirements.txt
openai
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
python-multipart